pyproject.toml
[tool.poetry]
name = "capsulewardrobe"
version = "0.1.0"
description = ""
authors = ["Monika <monika.belkaite@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
